requirements.txt
asgiref==3.11.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
django-filter==24.3
Pillow==12.1.0
sqlparse==0.5.5

# Production dependencies
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
dj-database-url==2.1.0
python-decouple==3.8
